requirements.txt
boto3>=1.34.0
kubernetes>=28.1.0
click>=8.1.0
pyyaml>=6.0
jinja2>=3.1.0
requests>=2.31.0
tabulate>=0.9.0
colorama>=0.4.6
packaging>=21.0